Refactor to extract dummy repository to the shared folder
