+++ spec/event_store_client/mapper/encrypted_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'support/dummy_repository'
+
 module EventStoreClient
   RSpec.describe Mapper::Encrypted do
     let(:data) do
@@ -38,27 +40,6 @@ module EventStoreClient
     describe '#deserialize' do
       # TODO
     end
-  end
-end
-
-class DummyRepository
-  class Key
-    attr_accessor :iv, :cipher, :id
-    def initialize(id:, **)
-      @id = id
-    end
-  end
-
-  def find(user_id)
-    Key.new(id: user_id)
-  end
-
-  def encrypt(*)
-    'darthvader'
-  end
-
-  def decrypt(*)
-    JSON.generate(first_name: 'Anakin', last_name: 'Skylwalker')
   end
 end
