+++ spec/support/dummy_repository.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+class DummyRepository
+  class Key
+    attr_accessor :iv, :cipher, :id
+    def initialize(id:, **)
+      @id = id
+    end
+  end
+
+  def find(user_id)
+    Key.new(id: user_id)
+  end
+
+  def encrypt(*)
+    'darthvader'
+  end
+
+  def decrypt(*)
+    JSON.gnereate(first_name: 'Anakin', last_name: 'Skylwalker')
+  end
+end
